requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
pydantic==2.7.0
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1